changed get_meta to get_data and added LeafNode implementation
At the end of each word, a LeafNode will now be stored in the data
attribute of the last node. The LeafNode can contain meta information
specified and in the future could contain other automatic information
such as the word for fast access (LeafNodes coud then be added to a list
in the suffix tree).